Big tomographic reconstructions working!

+++ libEM/emdata.cpp
@@ -423,8 +423,8 @@ void EMData::clip_inplace(const Region & area,const float& fill_value)
 		for (int j = 0; j < civ.y_iter; ++j) {
 
 			// Determine the memory increments as dependent on i and j
-			int dst_inc = dst_it_end - j*new_nx - i*new_sec_size;
-			int src_inc = src_it_end - j*prev_nx - i*prev_sec_size;
+			size_t dst_inc = dst_it_end - (size_t)j*new_nx - (size_t)i*new_sec_size;
+			size_t src_inc = src_it_end - (size_t)j*prev_nx - (size_t)i*prev_sec_size;
 			float* local_dst = rdata + dst_inc;
 			float* local_src = rdata + src_inc;
